# lib/rand: Enhance __fill_random_buf using the multiple random seed