@sungup 2022-08-09
Hi, @axboe, please help to review this PR.

Here is a PR for the __fill_random_buf function. Our team are testing the new
PCIe generation device, but master version FIO cannot reach the device
specification about the sequential write performance.

I have already reviewed the latest __fill_random_buf and random-fill-aes
branch, but there are some problem to improve that function

1. Latest version __fill_random_buf show nice performance than before, but it
   has not solve the instruction dependencies of MUL. So, that function can be
   more improved if reduce the MUL dependencies.
2. random-fill-aes function is very goot performance on the Intel systems. But
   AMD system cannot shows performance gain like Intel CPU, and aes feature is
   not supported on the ARM architecture.

Current master version use MUL instruction while generating random number, but
MUL is known as a multiple-cycle instruction and the latency is longer than
ADD/SUB instruction. I assumed that single random-seed is cause of CPU stall,
and I have changed the function uses multiple-random seed. Also to prove my
assumtion, I compared the performance between the current __fill_random_buf,
fill_random_buf_aes, and PR requested version. To compare each algorithm, I
create simple c/c++ main function and calls 8 Million times get enough result.

Tested systems are Intel Skylake, AMD Epyc CPU, Apple M1 Max, and Ampere ARM
CPU. Following table is the result of my benchmark.


| Elapsed Time(sec) | master |  random-fill-aes | 2buckets | 4buckets | 8buckets | 16buckets |
|:------------------|-------:|-----------------:|---------:|---------:|---------:|----------:|
| Intel Icelake     | 131.0  | 65.5             | 93.7     | 55.0     | 54.8     | 54.9      |
| AMD Epyc          | 125.2  | 83.8             | 63.0     | 48.0     | 52.6     | 42.3      |
| ARM (M1 Max)      | 128.9  | Not supported    | 68.4     | 36.6     | 25.8     | 23.1      |
| ARM (Ampere)      | 183.1  | Not supported    | 137.6    | 137.5    | 137.6    | 137.5     |

As a result, random-fill-aes shows improvement on Intel/AMD, but multiple-seed
using over 4 buckets shows better result than using aes. Also, M1 Max shows the
most reduced overhead using the multiple-seed buckets.