@axboe 2022-08-09
This is great, I did ponder the dependency when I did the aes work, but didn't do anything about it. This pull request is much appreciated! And I agree that it's nice to have the generic filler be as fast as it can be as well.

Any chance you'd want to play with the aes variant too, to kill the dependency?

I'm going to tag a new release very shortly, and I think it'd be prudent to pull this one after making the release.