@vincentkfu 2022-08-09
Hello Sungup, here are some comments:

- please use C-style comments (/* ... */ instead of //)
- If the number of buckets will be a build-time option, it would be better to have it as an option in the configure script and `echo "#define SEED_BUCKETS $VALUE" > $config_host_h` instead of having it configured via the Makefile.

I did some quick tests and found big improvements with your patch running commands similar to

```
./fio --name=test --ioengine=null --rw=write --bs=128k --size=100g --refill=1
```
Below are bandwidth measurements on Intel Skylake:

| block size | PR | fio-3.30-97-g6e44f |
| :- | -: | -: |
| 4K | 11800 MiB/s | 6572 MiB/s |
| 128K | 24100 MiB/s | 9572 MiB/s |