@sungup 2022-08-10
Hello @axboe and @vincentkfu 
Thank you for your reply about this PR. :-)

I have change the code following the comments from Vincent, but I'm not familiar the configure script.
Please check the changes is right.