@axboe 2022-08-10
I ran this on my laptop (m1 pro running linux), and I get:

4K: 5611MB/sec -> 12.1GB/sec
128k: 7620MB/sec -> 28.4GB/sec

with the default 4 buckets.

Let's rename SEED_BUCKETS to CONFIG_SEED_BUCKETS and let's please squash the two patches into a single patch. You can do that while doing the rename.

Once that's done, we can pull this in! Thanks for doing this work.