@sungup 2022-08-10
Thank you for review.

I have squashed the commits and push into a single patch.
This command based squashing is my first time, and it is very good experience time to me.

I'm so happy this PR is helpful to you!